feat(espower): extract and export traversal rules
