+++ lib/instrumentor.js
@@ -1,28 +1,15 @@
 'use strict';
 
 var estraverse = require('estraverse'),
+    syntax = estraverse.Syntax,
     escallmatch = require('escallmatch'),
     SourceMapConsumer = require('source-map').SourceMapConsumer,
     clone = require('clone'),
     AssertionVisitor = require('./assertion-visitor'),
     EspowerError = require('./espower-error'),
-    typeName = require('type-name'),
-    syntax = estraverse.Syntax,
-    supportedNodeTypes = [
-        syntax.Identifier,
-        syntax.MemberExpression,
-        syntax.CallExpression,
-        syntax.UnaryExpression,
-        syntax.BinaryExpression,
-        syntax.LogicalExpression,
-        syntax.AssignmentExpression,
-        syntax.ObjectExpression,
-        syntax.NewExpression,
-        syntax.ArrayExpression,
-        syntax.ConditionalExpression,
-        syntax.UpdateExpression,
-        syntax.Property
-    ];
+    toBeSkipped = require('./rules/to-be-skipped'),
+    toBeCaptured = require('./rules/to-be-captured'),
+    typeName = require('type-name');
 
 function Instrumentor (options) {
     verifyOptionPrerequisites(options);
@@ -99,61 +86,6 @@ Instrumentor.prototype.instrument = function (ast) {
 function isCalleeOfParent(currentNode, parentNode) {
     return (parentNode.type === syntax.CallExpression || parentNode.type === syntax.NewExpression) &&
         parentNode.callee === currentNode;
-}
-
-function toBeCaptured (currentNode) {
-    switch(currentNode.type) {
-    case syntax.Identifier:
-    case syntax.MemberExpression:
-    case syntax.CallExpression:
-    case syntax.UnaryExpression:
-    case syntax.BinaryExpression:
-    case syntax.LogicalExpression:
-    case syntax.AssignmentExpression:
-    case syntax.UpdateExpression:
-    case syntax.NewExpression:
-        return true;
-    default:
-        return false;
-    }
-}
-
-function toBeSkipped (currentNode, parentNode, currentPath) {
-    return !isSupportedNodeType(currentNode) ||
-        isLeftHandSideOfAssignment(parentNode, currentPath) ||
-        isObjectLiteralKey(parentNode, currentPath) ||
-        isUpdateExpression(parentNode) ||
-        isCallExpressionWithNonComputedMemberExpression(currentNode, parentNode, currentPath) ||
-        isTypeOfOrDeleteUnaryExpression(currentNode, parentNode, currentPath);
-}
-
-function isLeftHandSideOfAssignment(parentNode, currentPath) {
-    // Do not instrument left due to 'Invalid left-hand side in assignment'
-    return parentNode.type === syntax.AssignmentExpression && currentPath === 'left';
-}
-
-function isObjectLiteralKey(parentNode, currentPath) {
-    // Do not instrument Object literal key
-    return parentNode.type === syntax.Property && parentNode.kind === 'init' && currentPath === 'key';
-}
-
-function isUpdateExpression(parentNode) {
-    // Just wrap UpdateExpression, not digging in.
-    return parentNode.type === syntax.UpdateExpression;
-}
-
-function isCallExpressionWithNonComputedMemberExpression(currentNode, parentNode, currentPath) {
-    // Do not instrument non-computed property of MemberExpression within CallExpression.
-    return currentNode.type === syntax.Identifier && parentNode.type === syntax.MemberExpression && !parentNode.computed && currentPath === 'property';
-}
-
-function isTypeOfOrDeleteUnaryExpression(currentNode, parentNode, currentPath) {
-    // 'typeof Identifier' or 'delete Identifier' is not instrumented
-    return currentNode.type === syntax.Identifier && parentNode.type === syntax.UnaryExpression && (parentNode.operator === 'typeof' || parentNode.operator === 'delete') && currentPath === 'argument';
-}
-
-function isSupportedNodeType (node) {
-    return supportedNodeTypes.indexOf(node.type) !== -1;
 }
 
 function verifyAstPrerequisites (ast, options) {
